Use common schema
